+++ source/app/rendering/graphcomponentscene.h
@@ -91,9 +91,9 @@ private:
     NodeId _queuedTransitionNodeId;
 
     bool _beingRemoved = false;
-    int _componentSize = 0;
+    size_t _componentSize = 0;
 
-    int _numComponentsPriorToChange = 0;
+    size_t _numComponentsPriorToChange = 0;
 
     void updateRendererVisibility();
